pyproject.toml
[project]
name = "launchpad-ai-engine"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10, <=3.13"

dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=0.19.1",
]
